Multiplication result converted to larger type
CWE-190
CWE-192
CWE-197
CWE-681

To fix this safely, ensure the multiplication is performed in size_t by casting
at least one operand before multiplying. This preserves behavior for valid
inputs while preventing intermediate unsigned int overflow.

Best targeted fix in src/BarShader.cpp:

- At line 192 region, change (m_Height - y) * linelength to
  static_cast<size_t>(m_Height - y) * linelength.
- Optionally (and cleanly), keep linelength as unsigned and cast only at use
  site, minimizing functional changes.
- No new methods or dependencies are needed.
- No import changes are required for this specific cast-based fix.

+++ src/BarShader.cpp
@@ -191,7 +191,8 @@ void CBarShader::Draw( wxDC* dc, int iLeft, int iTop, bool bFlat )
 			std::memcpy(buf + idx, buf, idx);
 		}
 		if (y < m_Height) {
-			std::memcpy(buf + idx, buf, (m_Height - y) * linelength);
+			std::memcpy(buf + idx, buf,
+				static_cast<size_t>(m_Height - y) * linelength);
 		}
 	} else {
 		// draw rounded bar
